Remove supplimentary stack
This was included to avoid fat pointers on the stack, as it would
make stack manipulations harder - however the two item stack manipulations
are easy enough without the stack, and other manipulations can be done
in combination with the return stack.

This means that the simplified vm will be hardware compatiable with
a larger range of hardware, including existing stack cpus.

This should also simplify a hardware implimentation.